+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 [![Donate via PayPal](https://img.shields.io/badge/Donate-PayPal-blue.svg?style=flat-square)](https://www.paypal.me/techtoday) 
 
 <p>
+<b>Version 1.0.45</b> April 2022<br/>
+- NEW: Additional players now resumes the queue as well (previously, only the main player was doing so).<br/>
+</p>
+<p>
 <b>Version 1.0.44</b> April 2022<br/>
 - NEW: you can now adjust the additional player's volume, adapting it to the main sonos player volume. This is useful in case you've some recessed speakers, "speaking" too low or some too near speakers, "speaking" too high. You can adapt the volume in the config window or dinamically via msg input.<br/>
 - Updated the README.<br/>